requirements.txt
python-dotenv
requests
ckanapi
